+++ src/file-io/file-io.zig
@@ -72,7 +72,7 @@ const PaddingVars = struct {
 };
 
 pub fn getPaddingVars(source_files: []const []const u8, allocator: std.mem.Allocator) error{OutOfMemory}!PaddingVars {
-    var longest_operand: usize = 0;
+    var longest_operand: u64 = 0;
     for (source_files) |file| {
         if (file.len > longest_operand) longest_operand = file.len;
     }
@@ -92,7 +92,7 @@ fn zigStrToCStr(str: []const u8) [*c]u8 {
 
 // i know you hate global vars but idk if zig has static types //
 var dot_count: u8 = 0;
-pub fn printf_operation(src: []const u8, dest: []const u8, longest_src: usize, padding_str: [*c]const u8, comptime operation: []const u8) void {
+pub fn printf_operation(src: []const u8, dest: []const u8, longest_src: u64, padding_str: [*c]const u8, comptime operation: []const u8) void {
     if (dot_count < 3) dot_count += 1 else dot_count -= 2;
     // printf may as well be its own programming language kek //
     _ = std.c.printf("\"%s\" %.*s--[%s]--> \"%s\"%.*s\n", zigStrToCStr(src), longest_src - src.len, padding_str, zigStrToCStr(operation), zigStrToCStr(dest), dot_count, "...");
